refactor: finished header
