+++ src/app/globals.css
@@ -104,3 +104,10 @@ a {
   -o-transition: ease-in-out 0.2s;
   transition: ease-in-out 0.2s;
 }
+
+.three-trans {
+  -webkit-transition: ease-in-out 0.3s;
+  -moz-transition: ease-in-out 0.3s;
+  -o-transition: ease-in-out 0.3s;
+  transition: ease-in-out 0.3s;
+}
