+++ src/components/header/themeToggle/ThemeToggle.module.css
@@ -1,0 +1,61 @@
+.button {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  position: relative;
+  color: var(--gray);
+  background: none;
+  border: 0;
+}
+
+.button p {
+  letter-spacing: 0.2em;
+  text-transform: uppercase;
+  font-size: 0.75rem;
+}
+
+.button p:first-child {
+  margin-right: 1.25rem;
+}
+
+.button p:last-child {
+  margin-left: 1.25rem;
+}
+
+.circle {
+  border: 2px solid var(--gray);
+  border-radius: 50%;
+  position: relative;
+  transition: all ease-in-out 0.3;
+  width: 0.75rem;
+  height: 0.75rem;
+  z-index: 2;
+}
+
+.darkCircle {
+  transform: translateX(100%);
+  background-color: var(--gray);
+}
+
+.lightCircle {
+  transform: translateX(-100%);
+  background-color: var(--white);
+}
+
+.line {
+  position: absolute;
+  background-color: var(--gray);
+  width: 1rem;
+  height: 0.125rem;
+  border-radius: 0.75rem;
+  right: -0.125rem;
+  z-index: 1;
+}
+
+.lineDark {
+  transform: translateX(-453%);
+}
+
+.lineLight {
+  transform: translateX(-390%);
+}
